Refactor: split oversized files, enforce stricter SwiftLint
Tier 1 cleanup of the largest SwiftLint offenders so several relaxed
thresholds can return to (or near) their defaults. No behaviour change.

- Split Eyedroppers.swift (540 lines) into Eyedroppers.swift + a new
  Eyedropper.swift; move storage helpers and palette management into
  extensions. Decompose the 75-line start() into focused helpers.
- Extract AppDelegate.applicationDidFinishLaunching setup into named
  helpers and move the @IBAction forwarding shims into extensions.
- Split HelpView.swift (443 lines) into HelpView.swift + HelpRows.swift
  + HelpData.swift; convert the 3-member `formats` tuple to a struct.
- Tighten .swiftlint.yml: restore type_body_length, function_body_length,
  cyclomatic_complexity and large_tuple to defaults; keep line_length,
  identifier_name and file_length relaxed with rationale. Mark the
  URLSchemeHandler router as an explicit cyclomatic_complexity exception.

Both targets build and the test suite passes.
