+++ Pika/Services/Eyedroppers.swift
@@ -20,45 +20,6 @@ class Eyedroppers: ObservableObject {
         }
     }
 
-    // MARK: - History helpers
-
-    private var autoHistory: [ColorPair] {
-        get { Defaults[.palettes].first?.pairs ?? [] }
-        set {
-            var palettes = Defaults[.palettes]
-            guard !palettes.isEmpty else { return }
-            palettes[0].pairs = newValue
-            Defaults[.palettes] = palettes
-        }
-    }
-
-    private func pushUndo() {
-        var stack = Defaults[.undoStack]
-        stack.insert(autoHistory, at: 0)
-        if stack.count > ColorPair.maxHistory {
-            stack = Array(stack.prefix(ColorPair.maxHistory))
-        }
-        Defaults[.undoStack] = stack
-        Defaults[.redoStack] = []
-    }
-
-    private var activePairs: [ColorPair] {
-        let palettes = Defaults[.palettes]
-        let idx = Defaults[.activePaletteIndex]
-        guard idx >= 0, idx < palettes.count else { return autoHistory }
-        return palettes[idx].pairs
-    }
-
-    private var activeIndex: Int {
-        let pairs = activePairs
-        if let id = activeHistoryID,
-           let index = pairs.firstIndex(where: { $0.id == id })
-        {
-            return index
-        }
-        return 0
-    }
-
     func updateActiveEntry() {
         let fgHex = foreground.color.toHexString()
         let bgHex = background.color.toHexString()
@@ -226,11 +187,6 @@ class Eyedroppers: ObservableObject {
         }
     }
 
-    private var hasActiveSelection: Bool {
-        guard let id = activeHistoryID else { return false }
-        return activePairs.contains { $0.id == id }
-    }
-
     func navigatePrevious() {
         let pairs = activePairs
         guard !pairs.isEmpty else { return }
@@ -279,9 +235,57 @@ class Eyedroppers: ObservableObject {
         }
         lastAutoHistoryID = nil
     }
+}
 
-    // MARK: - Palette management
+// MARK: - Storage helpers
 
+private extension Eyedroppers {
+    var autoHistory: [ColorPair] {
+        get { Defaults[.palettes].first?.pairs ?? [] }
+        set {
+            var palettes = Defaults[.palettes]
+            guard !palettes.isEmpty else { return }
+            palettes[0].pairs = newValue
+            Defaults[.palettes] = palettes
+        }
+    }
+
+    func pushUndo() {
+        var stack = Defaults[.undoStack]
+        stack.insert(autoHistory, at: 0)
+        if stack.count > ColorPair.maxHistory {
+            stack = Array(stack.prefix(ColorPair.maxHistory))
+        }
+        Defaults[.undoStack] = stack
+        Defaults[.redoStack] = []
+    }
+
+    var activePairs: [ColorPair] {
+        let palettes = Defaults[.palettes]
+        let idx = Defaults[.activePaletteIndex]
+        guard idx >= 0, idx < palettes.count else { return autoHistory }
+        return palettes[idx].pairs
+    }
+
+    var activeIndex: Int {
+        let pairs = activePairs
+        if let id = activeHistoryID,
+           let index = pairs.firstIndex(where: { $0.id == id })
+        {
+            return index
+        }
+        return 0
+    }
+
+    var hasActiveSelection: Bool {
+        guard let id = activeHistoryID else { return false }
+        return activePairs.contains { $0.id == id }
+    }
+}
+
+// MARK: - Palette management
+
+extension Eyedroppers {
     func savePalette(name: String) {
         let current = ColorPair(
             id: UUID(),
@@ -338,203 +342,5 @@ class Eyedroppers: ObservableObject {
         palettes[paletteIndex].pairs.removeAll { $0.id == pairID }
         Defaults[.palettes] = palettes
         if activeHistoryID == pairID { activeHistoryID = nil }
-    }
-}
-
-class Eyedropper: ObservableObject {
-    enum Types: String, CustomStringConvertible {
-        case foreground
-        case background
-
-        var description: String {
-            switch self {
-            case .foreground: return PikaText.textColorForeground
-            case .background: return PikaText.textColorBackground
-            }
-        }
-
-        var copySelector: Selector {
-            switch self {
-            case .foreground: return #selector(AppDelegate.triggerCopyForeground)
-            case .background: return #selector(AppDelegate.triggerCopyBackground)
-            }
-        }
-
-        var pickSelector: Selector {
-            switch self {
-            case .foreground: return #selector(AppDelegate.triggerPickForeground)
-            case .background: return #selector(AppDelegate.triggerPickBackground)
-            }
-        }
-
-        var systemPickerSelector: Selector {
-            switch self {
-            case .foreground: return #selector(AppDelegate.triggerSystemPickerForeground)
-            case .background: return #selector(AppDelegate.triggerSystemPickerBackground)
-            }
-        }
-
-        var pickNotification: Notification.Name {
-            switch self {
-            case .foreground: return .triggerPickForeground
-            case .background: return .triggerPickBackground
-            }
-        }
-
-        var copyNotification: Notification.Name {
-            switch self {
-            case .foreground: return .triggerCopyForeground
-            case .background: return .triggerCopyBackground
-            }
-        }
-
-        var systemPickerNotification: Notification.Name {
-            switch self {
-            case .foreground: return .triggerSystemPickerForeground
-            case .background: return .triggerSystemPickerBackground
-            }
-        }
-    }
-
-    let type: Types
-    var forceShow = false
-    var pendingChainCommit = false
-
-    let colorNames: [ColorName] = loadColors()!
-    var closestVector: ClosestVector!
-
-    @objc @Published public var color: NSColor
-
-    private var overlayWindow = ColorPickOverlayWindow()
-
-    init(type: Types, color: NSColor) {
-        self.type = type
-        self.color = color.usingColorSpace(.sRGB) ?? color
-
-        // Load colors
-        closestVector = ClosestVector(colorNames.map { $0.color.toRGB8BitArray() })
-    }
-
-    func getClosestColor() -> String {
-        colorNames[closestVector.compare(color)].name
-    }
-
-    func set(_ selectedColor: NSColor) {
-        color = selectedColor.usingColorSpace(.sRGB) ?? selectedColor
-    }
-
-    @objc func colorDidChange(sender: AnyObject) {
-        if let picker = sender as? NSColorPanel {
-            guard let srgbColor = picker.color.usingColorSpace(.sRGB) else { return }
-            color = srgbColor
-            NotificationCenter.default.post(name: .systemColorChanged, object: nil)
-        }
-    }
-
-    func picker() {
-        let panel = NSColorPanel.shared
-        panel.showsAlpha = false
-        panel.title = "\(type.rawValue.capitalized)"
-        panel.titleVisibility = .visible
-        panel.setTarget(self)
-        panel.color = color
-        panel.mode = .RGB
-        panel.colorSpace = Defaults[.colorSpace]
-        panel.orderFrontRegardless()
-        panel.setAction(#selector(colorDidChange))
-        panel.isContinuous = true
-    }
-
-    func start(chainContrasting: Bool = false) {
-        if Defaults[.hidePikaWhilePicking] {
-            if NSApp.mainWindow?.isVisible == true {
-                forceShow = true
-            }
-            NSApp.sendAction(#selector(AppDelegate.hidePika), to: nil, from: nil)
-        }
-
-        DispatchQueue.main.asyncAfter(deadline: .now()) {
-            if Defaults[.appMode].usesPopover {
-                NSApp.activate(ignoringOtherApps: true)
-            }
-            let sampler = NSColorSampler()
-            sampler.show { selectedColor in
-
-                if let selectedColor = selectedColor {
-                    let normalizedColor = selectedColor.usingColorSpace(.sRGB) ?? selectedColor
-
-                    if Defaults[.showColorOverlay] {
-                        let colorText = normalizedColor.toFormat(
-                            format: Defaults[.colorFormat], style: Defaults[.copyFormat]
-                        )
-                        let cursorPosition = NSEvent.mouseLocation
-                        self.overlayWindow.show(
-                            colorText: colorText,
-                            pickedColor: normalizedColor,
-                            nearCursor: cursorPosition,
-                            duration: Defaults[.colorOverlayDuration]
-                        )
-                    }
-
-                    self.set(normalizedColor)
-
-                    if chainContrasting,
-                       self.type == .foreground,
-                       let appDelegate = NSApp.delegate as? AppDelegate
-                    {
-                        // Defer committing the foreground pick — we'll record once the
-                        // background is also picked (or the chained pick is cancelled).
-                        let background = appDelegate.eyedroppers.background
-                        background.pendingChainCommit = true
-                        // Don't bounce Pika back into view between picks: forward the
-                        // forceShow intent to the background pick instead.
-                        if self.forceShow {
-                            self.forceShow = false
-                            background.forceShow = true
-                        }
-                        let delay: Double = Defaults[.showColorOverlay] ? 0.4 : 0.05
-                        DispatchQueue.main.asyncAfter(deadline: .now() + delay) {
-                            background.start()
-                        }
-                    } else {
-                        self.pendingChainCommit = false
-                        NotificationCenter.default.post(name: .colorPicked, object: nil)
-
-                        if Defaults[.copyColorOnPick] {
-                            NSApp.sendAction(self.type.copySelector, to: nil, from: nil)
-                        } else if Defaults[.appMode].usesPopover {
-                            NSApp.sendAction(#selector(AppDelegate.showPopover), to: nil, from: nil)
-                        } else {
-                            NSApp.sendAction(#selector(AppDelegate.showPika), to: nil, from: nil)
-                        }
-                    }
-                } else if self.pendingChainCommit {
-                    // Chained background pick was cancelled; commit the foreground change
-                    // that was deferred so it isn't lost. self.type is .background here, so
-                    // route copy-on-pick to the foreground selector explicitly.
-                    self.pendingChainCommit = false
-                    NotificationCenter.default.post(name: .colorPicked, object: nil)
-                    if Defaults[.copyColorOnPick] {
-                        NSApp.sendAction(Eyedropper.Types.foreground.copySelector, to: nil, from: nil)
-                    } else if Defaults[.appMode].usesPopover {
-                        NSApp.sendAction(#selector(AppDelegate.showPopover), to: nil, from: nil)
-                    } else {
-                        NSApp.sendAction(#selector(AppDelegate.showPika), to: nil, from: nil)
-                    }
-                }
-
-                if self.forceShow {
-                    self.forceShow = false
-                    if !Defaults[.appMode].usesPopover {
-                        NSApp.sendAction(#selector(AppDelegate.showPika), to: nil, from: nil)
-                    }
-                }
-
-                let panel = NSColorPanel.shared
-                if panel.isVisible {
-                    self.picker()
-                }
-            }
-        }
     }
 }
